# feat(createRequestHtmlFragment): implemented circuit breaker