--- a/package.json
+++ b/package.json
@@ -94,6 +94,8 @@
     "isomorphic-fetch": "^2.2.1",
     "lean-intl": "^4.2.2",
     "matcher": "^2.1.0",
+    "opossum": "^5.0.0",
+    "opossum-prometheus": "^0.1.0",
     "pidusage": "^2.0.17",
     "prom-client": "^11.5.3",
     "prometheus-gc-stats": "^0.6.2",