<!--- Provide a general summary of your changes in the Title above -->

## Description
<!--- Describe your changes in detail -->
This implements a circuit breaker using opossum. The circuit opens when requests are taking > 1s _and_ event loop lag is > 30ms. When the circuit is open there is no SSR. The circuit will go into a half open state after 10s. At that point it will attempt the full request again and if there is success within the allotted time, the circuit will close.

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->
This provides significant performance improvements.

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, tests ran to see how -->
<!--- your change affects other areas of the code, etc. -->
In addition to the unit tests, I did some load testing locally. I ran a json server using `json-server http://jsonplaceholder.typicode.com/db -d 800 -p 1337`, then changed SSR Frank to make its requests to `http://localhost:1337/todos`. Once that was set up I called autocannon with `autocannon localhost:3000/healthy-frank/ssr-frank -c 20 -d 20 --headers "Accept-Language= en-USen;q=0.5" --headers "User-Agent= curl/7.5.0"`. Below are the results for master and for this branch.

### master

```
Running 20s test @ http://localhost:3000/healthy-frank/ssr-frank
20 connections

┌─────────┬────────┬────────┬─────────┬─────────┬───────────┬──────────┬────────────┐
│ Stat    │ 2.5%   │ 50%    │ 97.5%   │ 99%     │ Avg       │ Stdev    │ Max        │
├─────────┼────────┼────────┼─────────┼─────────┼───────────┼──────────┼────────────┤
│ Latency │ 814 ms │ 835 ms │ 1090 ms │ 1104 ms │ 863.87 ms │ 67.49 ms │ 1113.15 ms │
└─────────┴────────┴────────┴─────────┴─────────┴───────────┴──────────┴────────────┘
┌───────────┬─────┬──────┬─────────┬─────────┬─────────┬────────┬────────┐
│ Stat      │ 1%  │ 2.5% │ 50%     │ 97.5%   │ Avg     │ Stdev  │ Min    │
├───────────┼─────┼──────┼─────────┼─────────┼─────────┼────────┼────────┤
│ Req/Sec   │ 0   │ 0    │ 22      │ 38      │ 22.8    │ 6.89   │ 20     │
├───────────┼─────┼──────┼─────────┼─────────┼─────────┼────────┼────────┤
│ Bytes/Sec │ 0 B │ 0 B  │ 1.47 MB │ 2.55 MB │ 1.47 MB │ 546 kB │ 176 kB │
└───────────┴─────┴──────┴─────────┴─────────┴─────────┴────────┴────────┘

Req/Bytes counts sampled once per second.

456 requests in 20.09s, 29.4 MB read
```

### This branch

```
Running 20s test @ http://localhost:3000/healthy-frank/ssr-frank
20 connections

┌─────────┬───────┬───────┬────────┬────────┬──────────┬───────────┬────────────┐
│ Stat    │ 2.5%  │ 50%   │ 97.5%  │ 99%    │ Avg      │ Stdev     │ Max        │
├─────────┼───────┼───────┼────────┼────────┼──────────┼───────────┼────────────┤
│ Latency │ 29 ms │ 36 ms │ 844 ms │ 909 ms │ 65.82 ms │ 155.32 ms │ 1167.56 ms │
└─────────┴───────┴───────┴────────┴────────┴──────────┴───────────┴────────────┘
┌───────────┬─────────┬─────────┬──────┬─────────┬─────────┬─────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%  │ 97.5%   │ Avg     │ Stdev   │ Min     │
├───────────┼─────────┼─────────┼──────┼─────────┼─────────┼─────────┼─────────┤
│ Req/Sec   │ 1       │ 1       │ 334  │ 594     │ 299.36  │ 258.11  │ 1       │
├───────────┼─────────┼─────────┼──────┼─────────┼─────────┼─────────┼─────────┤
│ Bytes/Sec │ 67.4 kB │ 67.4 kB │ 4 MB │ 5.48 MB │ 3.32 MB │ 1.84 MB │ 67.3 kB │
└───────────┴─────────┴─────────┴──────┴─────────┴─────────┴─────────┴─────────┘

Req/Bytes counts sampled once per second.

6k requests in 20.09s, 66.3 MB read
```

## Types of Changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Documentation (adding or updating documentation)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [ ] My change requires a change to the documentation and I have updated the documentation accordingly.
- [ ] These changes should be applied to a maintenance branch.
- [ ] This change requires cross browser checks.
- [x] Performance tests should be ran against the server prior to merging.
- [ ] This change impacts caching for client browsers.
- [ ] This change impacts HTTP headers.
- [ ] This change adds additional environment variable requirements for One App users.
- [ ] I have added the Apache 2.0 license header to any new files created.

## What is the Impact to Developers Using One App?
<!--- Please describe how your changes impacts developers using One App. -->
No impact to developers other than SSR turning off in the cases that the circuit opens